livemap: - rename 'mappings_scaling_limit' to 'aes_zoomin_limit' - rename 'constants_scaling_limit' to 'const_zoomin_limit' - alpha scaling for lines/paths/segments - alpha scaling for polygon/rect borders - alpha scaling level is now limited by zoom-in limit (no more transparent micro objects) - aes and constant zoomin limits are not swapped anymore

+++ livemap/src/commonMain/kotlin/jetbrains/livemap/chart/Components.kt
@@ -17,7 +17,7 @@ class ChartElementComponent : EcsComponent {
     var sizeScalingRange: ClosedRange<Int>? = null
     var alphaScalingEnabled: Boolean = false
     var scalingSizeFactor: Double = 1.0
-    var scalingAlphaValue: Double? = null
+    var scalingAlphaValue: Int? = null
 }
 
 class TextSpecComponent : EcsComponent {
